fix(rulebuilder): add comments on props, clean up story
